+++ src/components/manage/Form/Form.jsx
@@ -202,6 +202,10 @@ class Form extends Component {
           ? formData[blocksLayoutFieldname].items[0]
           : null,
       placeholderProps: {},
+      testData: [
+        { external_link: 'https://ddg.gg' },
+        { external_link: 'https://wikipedia.org' },
+      ],
     };
     this.onChangeField = this.onChangeField.bind(this);
     this.onChangeBlock = this.onChangeBlock.bind(this);