+++ src/config/Widgets.jsx
@@ -15,6 +15,8 @@ import TextWidget from '@plone/volto/components/manage/Widgets/TextWidget';
 import TokenWidget from '@plone/volto/components/manage/Widgets/TokenWidget';
 import WysiwygWidget from '@plone/volto/components/manage/Widgets/WysiwygWidget';
 import ObjectBrowserWidget from '@plone/volto/components/manage/Widgets/ObjectBrowserWidget';
+import ObjectWidget from '@plone/volto/components/manage/Widgets/ObjectWidget';
+import ObjectListWidget from '@plone/volto/components/manage/Widgets/ObjectListWidget';
 
 export const DatetimeWidget = loadable(() =>
   import('@plone/volto/components/manage/Widgets/DatetimeWidget'),
@@ -33,6 +35,8 @@ export const widgetMapping = {
     datetime: DatetimeWidget,
     password: PasswordWidget,
     file: FileWidget,
+    object: ObjectWidget,
+    object_list: ObjectListWidget,
   },
   vocabulary: {
     'plone.app.vocabularies.Catalog': ObjectBrowserWidget, //ReferenceWidget,